refactor(kitchen pos) : NameCreationService -> ProductNameCreationService
