+++ src/main/java/kitchenpos/product/domain/model/ProductNameCreationService.java
@@ -1,22 +1,21 @@
-package kitchenpos.common.domain;
+package kitchenpos.product.domain.model;
 
-import java.util.Objects;
 import kitchenpos.common.application.PurgomalumClient;
 import org.springframework.stereotype.Service;
 
 @Service
-public class NameCreationService {
+public class ProductNameCreationService {
     private static final String NAME_VALIDATION_EXCEPTION = "이름에 비속어가 존재합니다. 비속어를 제외해주세요!";
     private final PurgomalumClient purgomalumClient;
 
-    public NameCreationService(PurgomalumClient purgomalumClient) {
+    public ProductNameCreationService(PurgomalumClient purgomalumClient) {
         this.purgomalumClient = purgomalumClient;
     }
 
-    public Name createName(String name) {
-        if (Objects.isNull(name) || purgomalumClient.containsProfanity(name)) {
+    public ProductName createName(String name) {
+        if (purgomalumClient.containsProfanity(name)) {
             throw new IllegalArgumentException(NAME_VALIDATION_EXCEPTION);
         }
-        return new Name(name);
+        return new ProductName(name);
     }
 }
